+++ tests/_commands/test_export_task.py
@@ -14,6 +14,7 @@ import torch
 from lightning_utilities.core.imports import RequirementCache
 
 import lightly_train
+from lightly_train._commands.export_task import OnnxPrecision
 
 from .. import helpers
 
@@ -64,15 +65,15 @@ def dinov2_vits14_eomt_checkpoint(tmp_path_factory: pytest.TempPathFactory) -> P
 
 
 onnx_export_testset = [
-    (1, 42, 154, False),
-    (1, 42, 154, False),
-    (2, 14, 14, False),
-    (3, 140, 280, True),
-    (4, 266, 28, True),
+    (1, 42, 154, OnnxPrecision.F32_TRUE),
+    (1, 42, 154, OnnxPrecision.F32_TRUE),
+    (2, 14, 14, OnnxPrecision.F32_TRUE),
+    (3, 140, 280, OnnxPrecision.F16_TRUE),
+    (4, 266, 28, OnnxPrecision.F16_TRUE),
 ]
 
 
-@pytest.mark.parametrize("batch_size,height,width,half", onnx_export_testset)
+@pytest.mark.parametrize("batch_size,height,width,precision", onnx_export_testset)
 @pytest.mark.skipif(
     sys.version_info < (3, 9),
     reason="Requires Python 3.9 or higher for image preprocessing.",
@@ -85,7 +86,7 @@ def test_onnx_export(
     batch_size: int,
     height: int,
     width: int,
-    half: bool,
+    precision: OnnxPrecision,
     dinov2_vits14_eomt_checkpoint: Path,
     tmp_path: Path,
 ) -> None:
@@ -99,7 +100,7 @@ def test_onnx_export(
     onnx_path = tmp_path / "model.onnx"
     validation_input = torch.randn(batch_size, 3, height, width, device="cpu")
     expected_outputs = model(validation_input)
-    expected_output_dtypes = [torch.int64, torch.float16 if half else torch.float32]
+    expected_output_dtypes = [torch.int64, precision.torch()]
     # We use  torch.testing.assert_close to check if the model outputs the same as when we run the exported
     # onnx file with onnxruntime. Unfortunately the default tolerances are too strict so we specify our own.
     rtol = 1e-2
@@ -111,7 +112,7 @@ def test_onnx_export(
         checkpoint=dinov2_vits14_eomt_checkpoint,
         height=height,
         width=width,
-        half=half,
+        precision=precision,
         batch_size=batch_size,
         overwrite=True,
     )
@@ -121,8 +122,7 @@ def test_onnx_export(
     onnx.checker.check_model(onnx_path, full_check=True)
 
     session = ort.InferenceSession(onnx_path, providers=["CPUExecutionProvider"])
-    if half:
-        validation_input = validation_input.half()
+    validation_input = validation_input.to(precision.torch())
     ort_in = {"input": validation_input.numpy()}
     ort_outputs = session.run(["masks", "logits"], ort_in)
     ort_outputs = [torch.from_numpy(y).cpu() for y in ort_outputs]